requirements.txt
Flask==2.3.3
python-dotenv==1.0.0
requests==2.31.0
Pillow==10.0.0
python-jose==3.3.0
Flask-Login==0.6.2
supabase==1.0.3
Werkzeug==2.3.7
